+++ packages/react-codemirror/playwright-ct.config.ts
@@ -45,7 +45,10 @@ export default defineConfig({
   testMatch: '*.spec.tsx',
 
   projects: [
-    { name: 'chromium', use: { ...devices['Desktop Chrome'] } },
+    {
+      name: 'chromium',
+      use: { ...devices['Desktop Chrome'] },
+    },
     {
       name: 'firefox',
       use: { ...devices['Desktop Firefox'] },